requirements.txt
ephem==3.7.6.0
Django==1.9.4
python-dateutil==2.5.0
